Add Journal Schedule
The journal schedule is a basic table that I use in my work journal to keep track of meetings and link meeting notes back to the daily plan etc.

I use the snippet in my "New Page" template for the work journal, so when a new file is created, the schedule is pre-populated based on the data in the `Data/schedule.json` file. Yes, we work in fortnightly sprints, so this would create the sprint start/end ceremonies in every week, but it's easier to remove what isn't needed than add what is. I might get around to making this work a bit cleverer in the future.